Corrected the example web.xml
The example web.xml mentions setting an `init-param` with the path to the application.conf file.
However, in the implementation of the class `io.ktor.server.servlet.ServletApplicationEngine` this setting is loaded from the servlet context (which contains only the `context-param`s; the `init-param`s are loaded from the servlet config).
The implementation also removes the prefix 'io.ktor' and then searches for a parameter with the name 'ktor.config'. That is why the correct `param-name` in the web.xml should be 'io.ktor.ktor.config'.

I created a PR on the documentation because I was not sure if this is a bug in the source code of https://github.com/ktorio/ktor/blob/master/ktor-server/ktor-server-servlet/src/io/ktor/server/servlet/ServletApplicationEngine.kt or a bug in the documentation. If the use of `init-param` was intentional then the source should be updated to load this parameter from the `servletConfig` property instead of the `servletContext` property.

Disclaimer: I tested this behavior on WebLogic Server Version 12.2.1.3.0; it is possible that Tomcat or Jetty behave differently.