Correct docs for fatJar generation with shadow
The correct plugin to generate a fatJar is 'shadow', not
'plugin-shadow'. The latter is supposed to be used for gradle plugins.
Also it runs way longer then the correct one (1:40 vs. 0:23 in my
build).

Also suggesting to use latest version 2.0.4 of the plugin.

See https://github.com/johnrengelman/shadow/issues/359#issuecomment-383538259.